repo: remove bevy dependency from buoy-physics
