+++ crates/buoy-physics/src/ideal_gas.rs
@@ -1,11 +1,6 @@
 //! Ideal gas equations.
-#![allow(dead_code)]
 
 use std::ops::{Div, Mul};
-
-use avian3d::math::Scalar;
-use bevy::{asset::Asset, prelude::*, reflect::TypePath};
-use serde::Deserialize;
 use uom::si::{
     f32::{
         ThermodynamicTemperature, Pressure, Mass, Volume,
@@ -23,10 +18,6 @@ use crate::{
     constants::{GAS_CONSTANT, STANDARD_GRAVITY},
     geometry::sphere_volume,
 };
-
-pub(crate) fn plugin(_app: &mut App) {
-    // nothing yet
-}
 
 /// Volume (m³) of an ideal gas from its temperature (K), pressure (Pa),
 /// mass (kg) and molar mass (kg/mol).
@@ -50,7 +41,7 @@ pub fn ideal_gas_density(
 }
 
 /// Molecular species of a gas.
-#[derive(Component, Debug, Clone, PartialEq)]
+#[derive(Debug, Clone, PartialEq)]
 pub struct GasSpecies {
     pub name: String,
     pub abbreviation: String,
@@ -98,31 +89,8 @@ impl Default for GasSpecies {
     }
 }
 
-#[derive(Deserialize, Debug, Asset, TypePath, Clone)]
-pub struct GasSpeciesConfig {
-    pub name: String,
-    pub abbreviation: String,
-    pub molar_mass: f32, // [kg/mol]
-}
-
-impl GasSpeciesConfig {
-    pub fn to_species(&self) -> GasSpecies {
-        GasSpecies {
-            name: self.name.clone(),
-            abbreviation: self.abbreviation.clone(),
-            molar_mass: MolarMass::new::<kilogram_per_mole>(self.molar_mass),
-        }
-    }
-}
-
-#[derive(Deserialize, Debug, Asset, TypePath, Clone)]
-pub struct GasPropertiesConfig {
-    pub gases: Vec<GasSpeciesConfig>,
-    // materials: Vec<MaterialConfig>, // can be added later
-}
-
 /// Properties of an ideal gas per unit mass.
-#[derive(Component, Default, Debug, Clone, PartialEq)]
+#[derive(Default, Debug, Clone, PartialEq)]
 pub struct IdealGas {
     pub species: GasSpecies,
     pub mass: Mass,
