+++ crates/buoy-runtime/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "buoy-runtime"
-version = "0.3.0"
+version = "0.4.0"
 description = "Runtime for the Buoy simulator"
 edition = { workspace = true }
 authors = { workspace = true }
@@ -16,7 +16,6 @@ buoy-physics = { path = "../buoy-physics" }
 bevy = { workspace = true }
 avian3d = { workspace = true }
 uom = { workspace = true }
-bevy_common_assets = { workspace = true }
 serde = { workspace = true }
 bevy_repl = { version = "0.3.0", features = ["derive"] }
