+++ crates/buoy-ui/Cargo.toml
@@ -30,7 +30,6 @@ bevy = { workspace = true, features = [
 ] }
 avian3d = { workspace = true, optional = true, features = ["bevy_picking"] }
 bevy_egui = { version = "0.35.0" }
-bevy_common_assets = { workspace = true }
 bevy-inspector-egui = { version = "0.32.0", optional = true, features = [
     "highlight_changes",
 ] }
@@ -38,7 +37,6 @@ bevy-inspector-egui = { version = "0.32.0", optional = true, features = [
 [features]
 default = ["dev"]
 dev = [
-    "buoy-physics/dev",
     "buoy-runtime/dev",
     "bevy/dynamic_linking",
     "avian3d/diagnostic_ui",
